requirements.txt
fastapi
uvicorn
streamlit
supabase
google-generativeai
python-dotenv
pydantic
requests
